requirements.txt
pandas
numpy
requests
psycopg2-binary
argparse

# venv\\Scripts\\activate